Add the Ingredients List activity, started via search function in Home activity
The Home activity now has a "Search ingredients" field and an "Add" button.
The search field is an EditText widget with an Add button. When pressed, the Ingredients List activity is started,
and the search query is (for now) appended to a TextView widget in that activity.

The Ingredients List category has a Find Recipe button (no functionality yet) and a Clear List button.
The Clear List button clears the TextView containing ingredients entered by the user.

* Added Up navigation to toolbar on most Activities
* Removed Search button from all toolbars

The toolbar search button is (temporarily) removed because its implementation is currently too complicated and redundant.
For the sake of simplicity, focus is placed on the singular search function via the Home and Ingredients List activities.